pyproject.toml
[project]
name = "money-stuff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "markdownify>=1.2.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "spacy[transformers]>=3.8.11",
    "spacy-transformers>=1.3.9",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "spacy-curated-transformers>=0.3.1",
    "ipywidgets>=8.1.8",
]
